Hjul med mana bar
Jeg mangler stadig at lave visuel ability select og evt. ult.

+++ HUD Magi/Assets/AbilityData.cs
@@ -6,5 +6,6 @@ using UnityEngine;
 public class AbilityData
 {
     public string abilityName;
-    public float manaCost;
+    public int manaCost;
+
 }
